Move slice module
