+++ py-polars/polars/_utils/slice.py
